Various warning clean-ups from IntelliJ
Split ValueNode into a separate class to avoid possible race-condition in class loading.

Modified ArrayPathToken to be super-class of ArrayIndex/ArraySlice token
because IntelliJ warned about null pointer exception possibilities.

Removed redundant boolean checks.

Removed use of StringBuffer -> StringBuilder.

Removed dangling ;
