+++ json-path/src/main/java/com/jayway/jsonpath/Criteria.java
@@ -28,7 +28,9 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 import static com.jayway.jsonpath.internal.Utils.notNull;
-
+import com.jayway.jsonpath.internal.filter.ValueNodes;
+import static com.jayway.jsonpath.internal.filter.ValueNodes.ValueListNode;
+import static com.jayway.jsonpath.internal.filter.ValueNodes.PredicateNode;
 
 /**
  *
@@ -225,7 +227,7 @@ public class Criteria implements Predicate {
     public Criteria in(Collection<?> c) {
         notNull(c, "collection can not be null");
         this.criteriaType = RelationalOperator.IN;
-        this.right = new ValueNode.ValueListNode(c);
+        this.right = new ValueListNode(c);
         return this;
     }
 
@@ -263,7 +265,7 @@ public class Criteria implements Predicate {
     public Criteria nin(Collection<?> c) {
         notNull(c, "collection can not be null");
         this.criteriaType = RelationalOperator.NIN;
-        this.right = new ValueNode.ValueListNode(c);
+        this.right = new ValueListNode(c);
         return this;
     }
 
@@ -290,7 +292,7 @@ public class Criteria implements Predicate {
     public Criteria subsetof(Collection<?> c) {
         notNull(c, "collection can not be null");
         this.criteriaType = RelationalOperator.SUBSETOF;
-        this.right = new ValueNode.ValueListNode(c);
+        this.right = new ValueListNode(c);
         return this;
     }
 
@@ -315,7 +317,7 @@ public class Criteria implements Predicate {
     public Criteria all(Collection<?> c) {
         notNull(c, "collection can not be null");
         this.criteriaType = RelationalOperator.ALL;
-        this.right = new ValueNode.ValueListNode(c);
+        this.right = new ValueListNode(c);
         return this;
     }
 
@@ -389,7 +391,7 @@ public class Criteria implements Predicate {
      */
     public Criteria empty(boolean empty) {
         this.criteriaType = RelationalOperator.EMPTY;
-        this.right = empty ? ValueNode.TRUE : ValueNode.FALSE;
+        this.right = empty ? ValueNodes.TRUE : ValueNodes.FALSE;
         return this;
     }
 
@@ -401,7 +403,7 @@ public class Criteria implements Predicate {
      */
     public Criteria matches(Predicate p) {
         this.criteriaType = RelationalOperator.MATCHES;
-        this.right = new ValueNode.PredicateNode(p);
+        this.right = new PredicateNode(p);
         return this;
     }
