+++ json-path/src/main/java/com/jayway/jsonpath/internal/PathRef.java
@@ -212,7 +212,7 @@ public abstract class PathRef implements Comparable<PathRef>  {
         public int compareTo(PathRef o) {
             if(o instanceof ArrayIndexPathRef){
                 ArrayIndexPathRef pf = (ArrayIndexPathRef) o;
-                return Integer.valueOf(pf.index).compareTo(this.index);
+                return Integer.compare(pf.index, this.index);
             }
             return super.compareTo(o);
         }
