+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/CompiledPath.java
@@ -97,7 +97,7 @@ public class CompiledPath implements Path {
         try {
             PathRef op = ctx.forUpdate() ?  PathRef.createRoot(rootDocument) : PathRef.NO_OP;
             root.evaluate("", op, document, ctx);
-        } catch (EvaluationAbortException abort){};
+        } catch (EvaluationAbortException abort) {}
 
         return ctx;
     }
