+++ json-path/src/main/java/com/jayway/jsonpath/spi/cache/CacheProvider.java
@@ -6,7 +6,6 @@ import static com.jayway.jsonpath.internal.Utils.notNull;
 
 public class CacheProvider {
     private static Cache cache;
-    private static boolean cachingEnabled;
 
     public static void setCache(Cache cache){
         notNull(cache, "Cache may not be null");
@@ -16,7 +15,6 @@ public class CacheProvider {
             } else {
                 CacheProvider.cache = cache;
             }
-            cachingEnabled = !(CacheProvider.cache instanceof NOOPCache);
         }
     }
